todo:
it looks like the $_SESSION timestamp is reading correctly from the db so the error is happening when it trys to re read this value from the db, maybe I have to format it as a time stamp or something, I will try to make sql querys to get the verbage correct then try to see where it is formating it wrong in zend.

that one function still says it does not exist in sql so I need to iron out the details on that. once that is done I should be done with the back end stuff

maybe I should put something in so that all people have to enter a number or something also I should use cookies to save the timestamp in case they close there browser and then re open it later

the loading gif has not returned ever though I moved it back to the correct slot, should figure out why